Backend/requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
python-dotenv
psycopg2-binary
google-cloud-firestore
